some: cleanup
