+++ benchmarking/src/app/query-benchmarking/query-benchmarking.tsx
@@ -2,7 +2,6 @@ import { useState } from 'react';
 import { TEST_QUERIES } from '../constants';
 import { useDBM } from '../hooks/dbm-context';
 import { useClassicEffect } from '../hooks/use-classic-effect';
-import { NativeAppEvent } from '../native-app/electron-contants';
 
 export const QueryBenchmarking = () => {
   const [output, setOutput] = useState<
@@ -28,34 +27,32 @@ export const QueryBenchmarking = () => {
           query: TEST_QUERIES[i],
           tables: [{ name: 'taxi' }],
           options: {
-            preQuery: async (tablesFileData) => {
-              for (const table of tablesFileData) {
-                console.log(
-                  'getFilePathsForTable in query benchmarking',
-                  table.tableName
-                );
-                const filePaths: string[] =
-                  await window.electron?.ipcRenderer.invoke(
-                    NativeAppEvent.GET_FILE_PATHS_FOR_TABLE,
-                    {
-                      tableName: table.tableName,
-                    }
-                  );
-
-                console.log('filePaths', filePaths);
-                if (!filePaths) {
-                  throw new Error('File paths not found');
-                }
-
-                await dbm.query(
-                  `CREATE TABLE IF NOT EXISTS ${
-                    table.tableName
-                  } AS SELECT * FROM read_parquet(['${filePaths.join(
-                    "','"
-                  )}']);`
-                );
-              }
-            },
+            // preQuery: async (tablesFileData) => {
+            //   for (const table of tablesFileData) {
+            //     console.log(
+            //       'getFilePathsForTable in query benchmarking',
+            //       table.tableName
+            //     );
+            //     const filePaths: string[] =
+            //       await window.electron?.ipcRenderer.invoke(
+            //         NativeAppEvent.GET_FILE_PATHS_FOR_TABLE,
+            //         {
+            //           tableName: table.tableName,
+            //         }
+            //       );
+            //     console.log('filePaths', filePaths);
+            //     if (!filePaths) {
+            //       throw new Error('File paths not found');
+            //     }
+            //     await dbm.query(
+            //       `CREATE TABLE IF NOT EXISTS ${
+            //         table.tableName
+            //       } AS SELECT * FROM read_parquet(['${filePaths.join(
+            //         "','"
+            //       )}']);`
+            //     );
+            //   }
+            // },
           },
         })
         .then((results) => {
