+++ electron-app/src/app/constants.ts
@@ -1,4 +1,5 @@
 export const RENDER_APP_PORT = 4204;
 export const RENDERER_APP_NAME = 'benchmarking-app';
 
+// Base directory for parquet files
 export const BASE_DIR = 'electron-app/src/data';
